% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{AdjustedRand}
\alias{AdjustedRand}
\title{AdjustedRand}
\usage{
AdjustedRand(trial.classification, test.classification)
}
\arguments{
\item{trial.classification}{One classification.}

\item{test.classification}{Another classification.}
}
\description{
Checks the similarity of two classifications: 1 indicates perfect, 0 indicates as if random.
}

