% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{ExternalIndices}
\alias{ExternalIndices}
\title{ExternalIndices}
\usage{
ExternalIndices(k, Classification1, Classification2)
}
\arguments{
\item{k}{Number of clusters.}

\item{Classification1}{One classification.}

\item{Classification2}{Another classification.}
}
\description{
Measures of classification quality.
}
