% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss-data.R
\docType{data}
\name{gss}
\alias{gss}
\title{Data from the General Social Survey (2014)}
\format{A data.frame with 3842 rows and six
categorical variables.  Each categorical variable has levels \code{"A GREAT DEAL"},
\code{"ONLY SOME"}, and \code{"HARDLY ANY"}.}
\usage{
data(gss)
}
\description{
A dataset containing a subset of the results of the 2014 General
Social Survey, which polls Americans on, among other things, their
confidence in public institutions.
}
\details{
\itemize{
\item \code{confinan} - Confidence in banks and financial institutions.
\item \code{conbus} - Confidence in major companies.
\item \code{coneduc} - Confidence in education.
\item \code{conlabor} - Confidence in organized labor.
\item \code{conmedic} - Confidence in medicine.
\item \code{conpress} - Confidence in the press.
}
}
\references{
\url{http://http://gss.norc.org/get-the-data/},
    \url{http://ssric.org/node/551}
}
