% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{predict.Kmeans}
\alias{predict.Kmeans}
\title{\code{predict.KMeans}}
\usage{
\method{predict}{Kmeans}(object, newdata = object$model, ...)
}
\arguments{
\item{object}{A \code{KMeans} object.}

\item{newdata}{Optionally, a data frame in which to look for variables with which to predict.
If omitted, the data used to fit the model is used.}

\item{...}{Additional arguments to pass to predict.lda.
The default is to predict \code{NA}.}

\item{na.action}{Function determining what should be done with missing values in \code{newdata}.
The default is to predict \code{NA}.}
}
\description{
Predicts cluster membership for the entire sample passed into the original analysis (including missing and filtered values).
}

