% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumofsquares.R
\name{ResidualSumOfSquares}
\alias{ResidualSumOfSquares}
\title{ResidualSumOfSquares}
\usage{
ResidualSumOfSquares(x, groups, weights)
}
\arguments{
\item{x}{A \link{data.frame} or \link{matrix}.}

\item{groups}{A variable indicating group membership. Either a \link{factor} or coerced to a \link{factor}.}

\item{weights}{The sampling or replication weights.}
}
\description{
ResidualSumOfSquares
}
\details{
Missing values are ignored.
}
