% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumofsquares.R
\name{TotalSumOfSquares}
\alias{TotalSumOfSquares}
\title{TotalSumOfSquares}
\usage{
TotalSumOfSquares(x, weights)
}
\arguments{
\item{x}{A \link{data.frame} or \link{matrix}.}

\item{weights}{The sampling or replication weights.}
}
\description{
TotalSumOfSquares
}
\details{
Missing values are ignored.
}
