% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{BatchKMeans}
\alias{BatchKMeans}
\title{BatchKMeans}
\usage{
BatchKMeans(x, centers, weights, iter.max, n.starts, seed = 1223)
}
\arguments{
\item{x}{A \code{\link{data.frame}}.}

\item{centers}{Either the number of clusters (e.g., 2), or a set of initial cluster
centers.}

\item{weights}{An optional vector of sampling weights, or, the name or, the
name of a variable in \code{data}. It may not be an expression.}

\item{iter.max}{The number of iterations of the algorithm.}

\item{n.starts}{The number of times to run the whole algorihtm.}

\item{seed}{The seed for the random number generator.}
}
\description{
Uses the batch method k-means method for forming clusters.
}
\details{
The batch method works by selecting initial cluster centers, allocating each observation
to the closest cluster, recomputing the cluster centers, and repeating these steps until the either the
residual sum of squares stops reducing, or, \code{iter.max} is exceeded.

The two novel features of this algorithm, relative to traditional k-means algorithms such as Hartigan and Wong (1979) are:
(1) The algorithms addresses weights. (2) The algorithm classifies cases that have incomplete data.

 The first starting point of the algorithm is computed as follows: (1) Missing values are removed. (2) If the data is weighted,
 a new sample is created via resampling. (3) The Hartigan-Wong algorithm is applied to the syntheatic sample. (4) The cluster
 centers identified via Hartigan-Wong are used as the initial clusters of the batch method.

 Where \code{n.starts} is greater than 1, or, there are less than 100 cases left after removing cases with
 incomplete data, the remaining start points are selected by: (1) identifying unique cases, and (2) sampling
 without replacement from amongst the unique cases.
}
\references{
Hartigan, J. A. and Wong, M. A. (1979). A K-means clustering algorithm.
Applied Statistics 28, 100-108.
}
