% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{CalinskiHarabasz}
\alias{CalinskiHarabasz}
\title{\code{CalinskiHarabasz}}
\usage{
CalinskiHarabasz(x, centers, clusters)
}
\arguments{
\item{centers}{The cluster centers.}

\item{clusters}{Cluster membershis by case}

\item{data}{A matrix containing the data.}
}
\description{
Statistic for evaluating the fit of a cluster analysis (higher is better).
KMeans Cluster Analysis.
}

