% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{KMeans}
\alias{KMeans}
\title{\code{KMeans}}
\usage{
KMeans(data = NULL, centers = 2, subset = NULL, weights = NULL,
  missing = "Use partial data", iter.max = 100, n.starts = 10,
  algorithm = "Batch", output = "Means", seed = 1223, binary = FALSE,
  show.labels = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{data}{A \code{\link{data.frame}}.}

\item{centers}{Either the number of clusters (e.g., 2), or a set of initial cluster
centers. Where the number of clusters is specified, or the algorithm is 'Bagging',
a random selection of rows of data is chosen  as the initial start points.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process, or, the name of a variable in \code{data}. It
may not be an expression. \code{subset} may not}

\item{weights}{An optional vector of sampling weights, or, the name or, the
name of a variable in \code{data}. It may not be an expression.}

\item{missing}{How missing data is to be treated in the regression. Options:
\code{"Error if missing data"},
\code{"Exclude cases with missing data"},
\code{"Use partial data"}. This is the default.
\code{"Imputation (replace missing values with estimates)"}.}

\item{iter.max}{The number of iterations of the algorithm to run.}

\item{n.starts}{The number of times the algorithm should be run, each time with a different
number of start points.}

\item{algorithm}{One of \code{"Hartigan-Wong"}, \code{"Forgy"}, \code{"Lloyd"}, \code{"MacQueen"},\code{"Batch"}, or
\code{"Bagging"}.}

\item{output}{The defaults is \code{"Means"}. A table that is better for exporting is \code{"Means table"}.}

\item{seed}{The random number seed used in imputation.}

\item{binary}{Makes categorical variables into indicator variables (otherwise their values are used).}

\item{show.labels}{Shows the variable labels, as opposed to the labels, in the outputs, where a
variables label is an attribute (e.g., attr(foo, "label")).}

\item{verbose}{Whether or not to show the verbose outputs to \code{bclust}. Defaults to false.}

\item{...}{Additional arguments to \code{bclust}.}
}
\description{
KMeans Cluster Analysis.
}
\details{
\code{"Bagging"} uses bagging in an attempt to find replicable custers.
By default, 10 bootstrap samples are created (using weights if provided), and k-mean
cluster analysis is used to find 20 clusters in each of these samples, and the complete-link
hiearchical clustering algorithm is then used to form the final clusters (Leisch 1999).
See \code{\link{bclust}} to see the names and descriptions of additional parameters.
After running \code{\link{bclust}}, cases are assigned to the most similar cluster.
}
\references{
Forgy, E. W. (1965) Cluster analysis of multivariate data: efficiency vs interpretability of classifications. Biometrics 21, 768-769.
Hartigan, J. A. and Wong, M. A. (1979). A K-means clustering algorithm. Applied Statistics 28, 100-108.
Leisch, Friedrich (1999)  Bagged clustering. Working Paper 51, SFB "Adaptive Information Systems and Modeling in Economics and Management Science", August 1999. http://epub.wu.ac.at/1272/ 1/document.pdf
Lloyd, S. P. (1957, 1982) Least squares quantization in PCM. Technical Note, Bell Laboratories. Published in 1982 in IEEE Transactions on Information Theory 28, 128-137.
MacQueen, J. (1967) Some methods for classification and analysis of multivariate observations. In Proceedings of the Fifth Berkeley Symposium on Mathematical Statistics and Probability, eds L. M. Le Cam & J. Neyman, 1, pp. 281-297. Berkeley, CA: University of California Press.
}

