% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{print.KMeans}
\alias{print.KMeans}
\title{print.KMeans}
\usage{
\method{print}{KMeans}(x, ...)
}
\arguments{
\item{x}{The \link{KMeans} object.}

\item{...}{Generic print arguments.}
}
\description{
print.KMeans
}
